# BackEnd/static/swagger.yaml
openapi: 3.0.0
info:
  title: MoveHive - API
  version: 1.0.0
  description: API para gerenciar usuários e esportes na aplicação MoveHive.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /usuario/{usuario_id}:
    get:
      summary: Visualizar perfil do usuário
      description: >
        Retorna os dados do perfil de um usuário específico.  
        Inclui informações básicas do usuário, total de posts, número de seguidores e número de pessoas que o usuário está seguindo.  
        Campos sensíveis como senha, lista completa de seguidores e seguindo não são retornados.
      tags:
        - Usuário
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: usuario_id
          required: true
          schema:
            type: string
          description: ID único do usuário no banco de dados.
      responses:
        '200':
          description: Perfil retornado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  nome:
                    type: string
                    example: "João Silva"
                  email:
                    type: string
                    example: "joao@email.com"
                  bio:
                    type: string
                    example: "Apaixonado por esportes e tecnologia."
                  total_posts:
                    type: integer
                    example: 12
                  seguidores_count:
                    type: integer
                    example: 150
                  seguindo_count:
                    type: integer
                    example: 80
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Usuário não encontrado!"
        '500':
          description: Erro interno no servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Ocorreu um erro interno ao processar o perfil."

# Esporte
...
  /evento/AdicionarEvento:
   post:
    tags:
      - Eventos
    security:
        - bearerAuth: []
    summary: Adiciona um novo evento esportivo
    requestBody:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              titulo:
                type: string
                description: "Título do evento. Ex: 'Partida de futebol no parque'"
              descricao:
                type: string
                description: "Descrição do evento. Ex: 'Venha jogar uma partida amistosa no parque Ibirapuera.'"
              esporte_nome:
                type: string
                description: "Nome do esporte. Ex: 'Futebol'"
              data_hora_str:
                type: string
                description: "Data e hora do evento no formato 'YYYY-MM-DD HH:MM'. Ex: '2025-08-05 16:30'"
              localizacao:
                type: string
                description: "Local do evento. Ex: 'Parque Ibirapuera, São Paulo'"
              max_participantes:
                type: integer
                description: "Número máximo de participantes. Ex: 10"
              torneio:
                type: boolean
                description: "Indica se é um torneio. Ex: false"
              premiacao:
                type: string
                description: "Premiação do evento. Ex: 'Medalhas para os três primeiros colocados'"
              privado:
                type: boolean
                description: "Indica se o evento é privado. Ex: true"
              observacoes:
                type: string
                description: "Observações adicionais. Ex: 'Levar colete'"
              arquivo_foto:
                type: string
                format: binary
                description: "Imagem do evento (formato .jpg ou .png)"
            required:
              - titulo
              - descricao
              - esporte_nome
              - data_hora_str
              - localizacao
              - max_participantes
              - torneio
              - privado
        application/json:  # Apenas para exibição do exemplo na UI
          example:
            titulo: "Partida de futebol no parque"
            descricao: "Venha jogar uma partida amistosa no parque Ibirapuera."
            esporte_nome: "Futebol"
            data_hora_str: "2025-08-05 16:30"
            localizacao: "Parque Ibirapuera, São Paulo"
            max_participantes: 10
            torneio: false
            premiacao: "Medalhas para os três primeiros colocados"
            privado: true
            observacoes: "Levar colete"
    responses:
      '201':
        description: Evento criado com sucesso
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  example: "evt_123abc"
                titulo:
                  type: string
                  example: "Partida de futebol no parque"
                descricao:
                  type: string
                  example: "Venha jogar uma partida amistosa no parque Ibirapuera."
                esporte_nome:
                  type: string
                  example: "Futebol"
                data_hora:
                  type: string
                  format: date-time
                  example: "2025-08-05T16:30:00Z"
                localizacao:
                  type: string
                  example: "Parque Ibirapuera, São Paulo"
                max_participantes:
                  type: integer
                  example: 10
                torneio:
                  type: boolean
                  example: false
                premiacao:
                  type: string
                  example: "Medalhas para os três primeiros colocados"
                privado:
                  type: boolean
                  example: true
                observacoes:
                  type: string
                  example: "Levar colete"
                imagem_url:
                  type: string
                  format: uri
                  example: "https://storage.googleapis.com/....."
      '400':
        description: Erro na requisição (campos inválidos ou obrigatórios não fornecidos)
        content:
          application/json:
            schema:
              type: object
              properties:
                erro:
                  type: string
                  example: "Todos os campos obrigatórios devem ser preenchidos."
  /evento/meusEventos:
   get:
    tags:
      - Eventos
    summary: Lista os eventos criados pelo usuário autenticado
    description: Retorna todos os eventos que foram criados pelo usuário logado (identificado via token).
    security:
      - bearerAuth: []  
    responses:
      '200':
        description: Lista de eventos do usuário
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    example: "evt_123abc"
                  titulo:
                    type: string
                    example: "Partida de futebol no parque"
                  descricao:
                    type: string
                    example: "Venha jogar uma partida amistosa no parque Ibirapuera."
                  esporte_nome:
                    type: string
                    example: "Futebol"
                  data_hora:
                    type: string
                    format: date-time
                    example: "2025-08-05T16:30:00Z"
                  localizacao:
                    type: string
                    example: "Parque Ibirapuera, São Paulo"
                  max_participantes:
                    type: integer
                    example: 10
                  torneio:
                    type: boolean
                    example: false
                  premiacao:
                    type: string
                    example: "Medalhas para os três primeiros colocados"
                  privado:
                    type: boolean
                    example: true
                  observacoes:
                    type: string
                    example: "Levar colete"
                  imagem_url:
                    type: string
                    format: uri
                    example: "https://meusite.com/imagens/eventos/evt_123abc.jpg"
      '401':
        description: Não autorizado – token ausente ou inválido
        content:
          application/json:
            schema:
              type: object
              properties:
                erro:
                  type: string
                  example: "Token de autenticação inválido ou ausente."
  /evento/editarEvento:
   put:
    tags:
      - Eventos
    summary: Edita um evento existente
    description: Permite que o usuário edite um evento que ele criou, enviando os novos dados via formulário e imagem opcional.
    security:
      - bearerAuth: []  
    requestBody:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              id:
                type: string
                description: "ID do evento a ser editado. Obrigatório."
                example: "evt_123abc"
              titulo:
                type: string
                description: "Novo título do evento"
                example: "Jogo de vôlei na praia"
              descricao:
                type: string
                description: "Nova descrição do evento"
                example: "Traga sua bola e venha jogar com a galera!"
              esporte_nome:
                type: string
                description: "Nome do esporte"
                example: "Vôlei"
              data_hora:
                type: string
                description: "Nova data e hora no formato 'YYYY-MM-DD HH:MM'"
                example: "2025-08-10 09:00"
              localizacao:
                type: string
                description: "Novo local do evento"
                example: "Praia de Copacabana, RJ"
              max_participantes:
                type: integer
                description: "Novo número máximo de participantes"
                example: 12
              torneio:
                type: boolean
                description: "Indica se é um torneio"
                example: false
              premiacao:
                type: string
                description: "Premiação (caso torneio)"
                example: "Camisas oficiais para os vencedores"
              privado:
                type: boolean
                description: "Indica se o evento é privado"
                example: true
              observacoes:
                type: string
                description: "Observações adicionais"
                example: "Tragam água"
              arquivo_foto:
                type: string
                format: binary
                description: "Imagem atualizada do evento (opcional)"
            required:
              - id
    responses:
      '200':
        description: Evento editado com sucesso
        content:
          application/json:
            schema:
              type: object
              properties:
                mensagem:
                  type: string
                  example: "Evento atualizado com sucesso."
                evento_id:
                  type: string
                  example: "evt_123abc"     
      '404':
        description: Evento não encontrado
        content:
          application/json:
            schema:
              type: object
              properties:
                erro:
                  type: string
                  example: "Evento não encontrado."  

      '400':
        description: "Requisição inválida (ex: ID ausente ou dados mal formatados)"
        content:
          application/json:
            schema:
              type: object
              properties:
                erro:
                  type: string
                  example: "O campo 'id' é obrigatório." 
  /evento/deletarEvento:
    delete:
      tags:
        - Eventos
      summary: Deleta um evento pelo ID
      description: Deleta o evento especificado pelo `evento_id` enviado no corpo da requisição JSON.
      security:
        - bearerAuth: [] 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                evento_id:
                  type: string
                  description: "ID do evento a ser deletado"
                  example: "evt_123abc"
              required:
                - evento_id
      responses:
        '200':
          description: Evento deletado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Evento deletado com sucesso."
                  evento_id:
                    type: string
                    example: "evt_123abc"
        '400':
          description: "Requisição inválida (ex: evento_id ausente)"
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "O campo 'evento_id' é obrigatório."
        '404':
          description: Evento não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Evento não encontrado."
  /evento/listarEventos:
    get:
      tags:
        - Eventos
      summary: Lista todos eventos.
      description: >
        Retorna uma lista de todos os eventos que o campo Torneio == False.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de eventos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "a7b93e90-f5d6-4f0e-b5a0-8c34c77a9a00"
                    usuario_id:
                      type: string
                      example: "user_123"
                    titulo:
                      type: string
                      example: "Campeonato de Vôlei de Praia"
                    descricao:
                      type: string
                      example: "Torneio aberto para equipes amadoras."
                    esporte_nome:
                      type: string
                      example: "Vôlei de Praia"
                    localizacao:
                      type: string
                      example: "Praia de Copacabana, Rio de Janeiro"
                    data_hora:
                      type: string
                      format: date-time
                      example: "2025-08-10T14:30:00Z"
                    max_participantes:
                      type: integer
                      example: 16
                    data_criacao:
                      type: string
                      format: date-time
                      example: "2025-07-28T12:00:00Z"
                    torneio:
                      type: boolean
                      example: true
                    premiacao:
                      type: string
                      example: "Medalhas e troféu"
                    foto:
                      type: string
                      example: "https://firebasestorage.googleapis.com/v0/b/app-eventos/o/evento123.jpg"
                    participantes:
                      type: array
                      items:
                        type: string
                      example: ["user_123", "user_456", "user_789"]
                    privado:
                      type: boolean
                      example: false
                    observacoes:
                      type: string
                      example: "Chegar com 15 minutos de antecedência"
                    status:
                      type: string
                      example: "Inscricoes_abertas"
        '401':
          description: Token inválido ou ausente
        '500':
          description: Erro interno do servidor
  /evento/listarTorneios:
    get:
      tags:
        - Eventos
      summary: Lista os torneios 
      description: >
        Retorna uma lista de eventos que o campo "Torneio" == True.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de torneios retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "c9e6b114-baf6-4552-bb11-b60f4bbdd6fc"
                    usuario_id:
                      type: string
                      example: "user_456"
                    titulo:
                      type: string
                      example: "Torneio Municipal de Futebol"
                    descricao:
                      type: string
                      example: "Torneio anual de futebol amador com premiação em dinheiro."
                    esporte_nome:
                      type: string
                      example: "Futebol"
                    localizacao:
                      type: string
                      example: "Estádio Municipal de São Caetano"
                    data_hora:
                      type: string
                      format: date-time
                      example: "2025-09-05T10:00:00Z"
                    max_participantes:
                      type: integer
                      example: 20
                    data_criacao:
                      type: string
                      format: date-time
                      example: "2025-07-15T09:45:00Z"
                    torneio:
                      type: boolean
                      example: true
                    premiacao:
                      type: string
                      example: "R$ 5.000 + troféus"
                    foto:
                      type: string
                      example: "https://firebasestorage.googleapis.com/v0/b/app-eventos/o/torneio_fut.jpg"
                    participantes:
                      type: array
                      items:
                        type: string
                      example: ["user_111", "user_222", "user_333"]
                    privado:
                      type: boolean
                      example: false
                    observacoes:
                      type: string
                      example: "Levar chuteira e documento com foto"
                    status:
                      type: string
                      example: "ativo"
        '401':
          description: Token inválido ou ausente
        '500':
          description: Erro interno do servidor
  /evento/participarEvento:
    post:
      tags:
        - Eventos
      summary: Participa de um evento/torneio
      description: >
        Permite que um usuário autenticado participe de um evento, desde que:
        - Não seja o criador do evento;
        - Não esteja já participando;
        - Ainda haja vagas disponíveis.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - evento_id
              properties:
                evento_id:
                  type: string
                  description: ID do evento que o usuário deseja participar
                  example: "d2a6f3c9-8273-4f0f-9340-b8b123abc123"
      responses:
        '200':
          description: Participação confirmada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Participação confirmada com sucesso!"
        '400':
          description: Erro de validação (usuário já participa, é o criador ou evento está cheio)
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Você já está participando deste evento."
        '404':
          description: Evento não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Evento não encontrado."
        '500':
          description: Erro interno ao tentar participar do evento
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Erro ao participar do evento: erro inesperado."
  /evento/cancelarParticipacao:
    post:
      tags:
        - Eventos
      summary: Cancela a participação do usuário em um evento
      description: >
        Permite que um usuário autenticado cancele sua participação em um evento, desde que:
        - Ele esteja participando;
        - Ele **não** seja o criador do evento.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - evento_id
              properties:
                evento_id:
                  type: string
                  description: ID do evento do qual o usuário deseja sair
                  example: "c4e7d890-91ad-44c7-a1ef-fae8c7751e0f"
      responses:
        '200':
          description: Participação cancelada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Participação cancelada com sucesso."
        '400':
          description: Erro de validação (usuário não participa ou é o criador do evento)
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Você não está participando deste evento."
        '404':
          description: Evento não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Evento não encontrado."
        '500':
          description: Erro interno ao tentar cancelar a participação
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Erro ao cancelar participação: erro inesperado."
  /evento/usuario/{usuario_id}:
     get:
      summary: Listar eventos de um usuário
      description: >
        Retorna todos os eventos cadastrados por um usuário específico, com base no seu **usuario_id**.  
        Caso o usuário não tenha eventos, retorna uma mensagem informando.
      tags:
        - Eventos
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: usuario_id
          required: true
          schema:
            type: string
          description: ID único do usuário no banco de dados.
      responses:
        '200':
          description: Lista de eventos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "abc123xyz"
                    titulo:
                      type: string
                      example: "Campeonato de Futebol"
                    descricao:
                      type: string
                      example: "Evento esportivo para amantes de futebol."
                    data:
                      type: string
                      format: date
                      example: "2025-09-15"
                    local:
                      type: string
                      example: "São Paulo - SP"
                    usuario_id:
                      type: string
                      example: "user_123"
        '404':
          description: Nenhum evento encontrado para este usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Nenhum evento encontrado para este usuário."
        '500':
          description: Erro interno no servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    example: "Erro inesperado ao buscar eventos."
  
# Posts 
...